refactor: rename AbstractMultifeatureDataclass for consistency
- Renamed `abstract_multi_trajectory_dataclass` to `abstract_multifeature_dataclass`
  across the codebase to improve naming consistency and clarify intent.
- Updated imports, type annotations, and references in related modules and tests
  to reflect the renaming.
- Refactored associated test assertions and maintained compatibility with the
  existing functionality.

Issue TCT-52
