+++ src/trajectory_container_tools/dataclasses/core/__init__.py
@@ -1,5 +1,5 @@
 # coding=utf-8
-from .abstract_multi_trajectory_dataclass import AbstractMultifeatureDataclass
+from .abstract_multifeature_dataclass import AbstractMultifeatureDataclass
 from .abstract_no_trajectory_dataclass import AbstractNoTrajectoryDataclass
 from .abstract_trajectory_dataclass import AbstractTrajectoryDataclass
 from .base_trajectory_dataclass import (
